# Removing unsafe parts of Decode method in QueryStringEnumerable