--- a/src/Shared/QueryStringEnumerable.cs
+++ b/src/Shared/QueryStringEnumerable.cs
@@ -57,8 +57,6 @@ internal
     /// </summary>
     public readonly struct EncodedNameValuePair
     {
-        private static readonly SearchValues<char> DecodingChars = SearchValues.Create("%+");
-
         /// <summary>
         /// Gets the name from this name/value pair in its original encoded form.
         /// To get the decoded string, call <see cref="DecodeName"/>.
@@ -96,7 +94,7 @@ internal
             // If the value is short, it's cheap to check up front if it really needs decoding. If it doesn't,
             // then we can save some allocations.
             ReadOnlySpan<char> source = chars.Span;
-            if (source.Length < 16 && source.IndexOfAny(DecodingChars) < 0)
+            if (source.Length < 16 && source.IndexOfAny('%', '+') < 0)
             {
                 return chars;
             }